reorg ops
