+++ package.json
@@ -58,8 +58,8 @@
     "echidna": " echidna-test contracts/Echidna.sol --contract Echidna --crytic-args=\"--solc-remap @rari-capital/=node_modules/@rari-capital/\" --config ops/echidna.conf.yaml",
     "deploy-etheno": "hardhat clean && hardhat compile && bash ops/deploy.sh",
     "console": "bash ops/console.sh",
-    "start": "bash ops/start.sh",
-    "restart": "docker container stop consideration_ethprovider && bash ops/start.sh",
+    "start": "bash ops/start-console.sh",
+    "restart": "docker container stop consideration_ethprovider && bash ops/start-console.sh",
     "stop": "docker container stop consideration_ethprovider",
     "test": "hardhat test",
     "profile": "REPORT_GAS=true hardhat test",
